rename tests.unit.modules.selenium
