+++ .travis.yml
@@ -29,7 +29,7 @@ jobs:
     - PYTHON=python3
     - PIP_INSTALL="sudo pip3 install"
     - RVM_KEY="409B6B1796C275462A1703113804BB82D39DC0E3"
-    - OSX_EXCLUDE="--exclude-test=tests.unit.modules.selenium.test_csharp"
+    - OSX_EXCLUDE="--exclude-test=tests.unit.modules._selenium.test_csharp"
 
 addons:
   firefox: latest
