+++ tests/unit/modules/_selenium/test_selenium_executor.py
@@ -16,7 +16,7 @@ from bzt.modules.selenium import SeleniumExecutor
 from bzt.utils import LDJSONReader, FileReader, is_windows
 from tests.unit import BZTestCase, RESOURCES_DIR, ROOT_LOGGER, EngineEmul
 from tests.unit.mocks import DummyListener
-from tests.unit.modules.selenium import SeleniumTestCase
+from tests.unit.modules._selenium import SeleniumTestCase
 
 
 class LDJSONReaderEmul(object):
